+++ include/TGUI/String.hpp
@@ -1430,8 +1430,10 @@ namespace tgui
 
 #if TGUI_COMPILED_WITH_CPP_VER >= 17
     using StringView = std::u32string_view;
+    using CharStringView = std::string_stream;
 #else
     using StringView = String;
+    using CharStringView = std::string;
 #endif
 }