+++ src/Color.cpp
@@ -46,6 +46,9 @@ namespace tgui
 #if defined(__GNUC__)
     #pragma GCC diagnostic push
     #pragma GCC diagnostic ignored "-Wdeprecated-declarations"
+#elif defined (_MSC_VER)
+    #pragma warning(push)
+    #pragma warning(disable: 4996)
 #endif
     const std::map<String, Color> Color::colorMap =
     {
@@ -61,6 +64,8 @@ namespace tgui
     };
 #if defined(__GNUC__)
     #pragma GCC diagnostic pop
+#elif defined (_MSC_VER)
+    #pragma warning(pop)
 #endif
 
 #if TGUI_COMPILED_WITH_CPP_VER < 17